requirements.txt
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
distro==1.9.0
Django==5.2.7
django-apscheduler==0.7.0
django-cors-headers==4.9.0
django-crontab==0.7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
groq==0.35.0
grpcio==1.76.0
grpcio-status==1.71.2
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
mysql-connector-python==9.5.0
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
pyswisseph==2.10.3.2
python-decouple==3.8
razorpay==2.0.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
srt==3.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
